+++ foundry/forms.py
@@ -40,7 +40,7 @@ class RememberMeCheckboxInput(forms.widgets.CheckboxInput):
 
     def render(self, *args, **kwargs):
         result = super(RememberMeCheckboxInput, self).render(*args, **kwargs)
-        return result + "Remember me"
+        return result + ugettext("Remember me")
 
 
 class SMSOptInCheckboxInput(forms.widgets.CheckboxInput):
@@ -58,6 +58,7 @@ class EmailOptInCheckboxInput(forms.widgets.CheckboxInput):
 
 
 class LoginForm(AuthenticationForm):
+    remember_me = forms.BooleanField(required=False, initial=True, label="", widget=RememberMeCheckboxInput)
 
     def __init__(self, *args, **kwargs):
         super(LoginForm, self).__init__(*args, **kwargs)
@@ -115,6 +116,7 @@ Note that both fields are case-sensitive."
 class JoinForm(UserCreationForm):
     """Custom join form"""
     accept_terms = forms.BooleanField(required=True, label="", widget=TermsCheckboxInput)
+    remember_me = forms.BooleanField(required=False, initial=True, label="", widget=RememberMeCheckboxInput)
 
     class Meta:
         model = models.Member
@@ -184,7 +186,7 @@ Please supply a different %(pretty_name)s." % {'pretty_name': pretty_name}
                 display_fields.append('dob')
         for name, field in self.fields.items():
             # Skip over protected fields
-            if name in ('id', 'username', 'password1', 'password2', 'accept_terms'):
+            if name in ('id', 'username', 'password1', 'password2', 'accept_terms', 'remember_me'):
                 continue
             if name not in display_fields:
                 del self.fields[name]
@@ -274,23 +276,55 @@ class EditProfileForm(forms.ModelForm):
 
     class Meta:
         model = models.Member
-        fields = ('email', 'image', 'dob', 'about_me', )
+        widgets = {'dob': OldSchoolDateWidget}
         
     def __init__(self, *args, **kwargs):
         self.base_fields['image'].widget = forms.FileInput()
-        self.base_fields['dob'].help_text = _("yyyy-mm-dd")
         super(EditProfileForm, self).__init__(*args, **kwargs)
+
+        # todo: need a preference member_edit_fields
+        display_fields = preferences.RegistrationPreferences.display_fields
+        for extra in ('image', 'dob', 'about_me'):
+            if extra not in display_fields:
+                display_fields.append(extra)
+        for name, field in self.fields.items():
+            # Skip over protected fields
+            if name in ('id',):
+                continue
+            if name not in display_fields:
+                del self.fields[name]
+
+        # Set some fields required
+        required_fields = preferences.RegistrationPreferences.required_fields
+        for name in required_fields:
+            field = self.fields.get(name, None)
+            if field and not field.required:
+                field.required = True
+
+    def clean(self):
+        cleaned_data = super(EditProfileForm, self).clean()
         
-    def clean_email(self):
-        # xxx: not necessarily required. Depends on preferences.
-        if self.cleaned_data['email'] and not self.cleaned_data['email'] == self.instance.email:
-            try:
-                User.objects.get(email=self.cleaned_data['email'])
-                raise forms.ValidationError(_('This email already has an account linked to it. Please use another.'))
-            except User.DoesNotExist:
-                return self.cleaned_data['email']
-        else:
-            return self.cleaned_data['email']
+        # Validate required fields
+        required_fields = preferences.RegistrationPreferences.required_fields
+        for name in required_fields:
+            value = self.cleaned_data.get(name, None)
+            if not value:
+                message = _("This field is required.")
+
+        # Validate unique fields
+        unique_fields = preferences.RegistrationPreferences.unique_fields
+        for name in unique_fields:
+            value = self.cleaned_data.get(name, None)
+            if value is not None:
+                di = {'%s__iexact' % name:value}
+                if models.Member.objects.filter(**di).count() > 0:
+                    pretty_name = self.fields[name].label.lower()
+                    message =_("The %(pretty_name)s is already in use. \
+Please supply a different %(pretty_name)s." % {'pretty_name': pretty_name}
+                    )
+                    self._errors[name] = self.error_class([message])
+
+        return cleaned_data
 
     as_div = as_div
 
@@ -365,7 +399,7 @@ class PasswordResetForm(BasePasswordResetForm):
 class AgeGatewayForm(forms.Form):
     country = forms.ModelChoiceField(queryset=models.Country.objects.all())
     date_of_birth = forms.DateField(widget=OldSchoolDateWidget)
-    remember_me = forms.BooleanField(required=False, label="", widget=RememberMeCheckboxInput)
+    remember_me = forms.BooleanField(required=False, initial=True, label="", widget=RememberMeCheckboxInput)
 
     def clean(self):
         cleaned_data = super(AgeGatewayForm, self).clean()